Add files via upload
Here is a quick map visualising what I did in the GIS software I use (QGIS). Not really sure how else to share the work. Basically what I did was took the geometric files of all the census tracts, selected the ones that were within the counties we were interested in. Computed centroid for each census tract. Imported the geometric files of the US coast from governmental data. Computed the shortest distance between each centroid and the line of the coast. Then I was able to export the csv with the distance to the coast from the centroid of each census tract. 

Let me know if everything is clear,
cheers, Michal.